Composer: Add `static-review` dependency
This pulls in a fork of the original package. This is because we have
dependencies such as Symfony Process package, version 4.0. The original
`static-review` only supports version 3.0.

The idea will be to use some of the features from this package, but in
the future we will hope to remove any features used from this package,
and use our own code.

This is because the `static-review` package is now abandoned.