Add `GitFilesFinder` class
This will be used to be able to filter a list of changed files, by a
specified list of particular file names or patterns.

The reason the Regex third party package has been used here, is so that
in the `git-review.yml.dist` file we can specify a list of paths that we
can limit certain tools like the ESLint checker using wildcard
directories.

E.g

```
paths:
  - "example/*/tests"
```

Etc...