Add `es-lint` command
This command will be used to run the ESLint tool on only the changed
files on a topic branch. This allows developers to focus on the things
they are working on directly.

If the current checked out branch is `master`, then the ESLint command
will run on all of the paths that can be specified in the
`git-review.yml.dist` configuration file.

The `.php_cs.dist` has been updated so that the `tests/fixtures` folder
is not scanned for php violations.