Add `php-cs-fixer` command
This command will be used to run the Php-CS-Fixer tool on only
the changed files on a topic branch. This allows developers to focus
on the things they are working on directly.

If the current checked out branch is `master`, then the PHP-CS-Fixer
command will run on all of the paths that can be specified in the
`git-review.yml.dist` configuration file.