Ensure `git-review` command is executable
When trying to run `./git-review` to display the available commands and
options, we receive the following error:

```

In Container.php line 948:

  Unresolvable dependency resolving [Parameter #0 [ <required>
$commandline ]] in class Shopworks\Git\Review\Process\Process

```

An integration test has been added, to at least prove that the
`./git-review` command is at least instantiable. This also helps to see
that the appropriate commands are registered.